D13_CallBoard_Project/requirements.txt
Django==3.2.2
django-allauth==0.44.0
django-apscheduler==0.5.2
django-filter==2.4.0
django-cleanup==5.2.0
